adding express router for simple db record stats
