+++ server/queries/latestRecord.sql
@@ -1,2 +1,4 @@
-SELECT min(time) as first_time, max(time) as latest_time FROM binance_klines
-WHERE exchange == X and symbol == Y;
+SELECT max(time) as latest FROM binance_klines WHERE exchange == X and symbol == Y;
+
+SELECT exchange, symbol, count(*), min(time) as earliest, max(time) as latest 
+FROM binance_klines GROUP BY exchange, symbol;
